Split monolithic pic app into separate model and view files
This is to allow code sharing with the upcoming logic gate activity.
