+++ app/src/models/pic/led.js
@@ -1,0 +1,144 @@
+var LEDView = React.createFactory(require('../../views/pic/led')),
+    Pin = require('./pin'),
+    Segment = require('./segment'),
+    constants = require('../../views/pic/constants'),
+    layout = require('../../views/pic/layout');
+
+var LED = function () {
+  var i, pin, segmentLayoutMap, segment;
+
+  this.name = 'led';
+  this.view = LEDView;
+
+  this.pins = [];
+  this.pinMap = {};
+  for (i = 0; i < 10; i++) {
+    pin = {
+      number: i,
+      inputMode: true,
+      placement: i < 5 ? 'top' : 'bottom',
+      x: 0,
+      y: 0,
+      height: 0,
+      width: 0,
+      labelSize: 0,
+      component: this,
+      notConnectable: [2, 7].indexOf(i) !== -1
+    };
+    pin.label = {
+      x: 0,
+      y: 0,
+      anchor: 'end',
+      text: ['g', 'f', 'ca', 'a', 'b', 'e', 'd', 'ca', 'c', 'DP'][i]
+    };
+    pin = new Pin(pin);
+    this.pins.push(pin);
+    this.pinMap[pin.label.text] = pin;
+  }
+
+  this.segments = [];
+  segmentLayoutMap = [
+    {x: 0, y: 0, rotation: 0},
+    {x: 1, y: 0, rotation: 90},
+    {x: 1, y: 1, rotation: 90},
+    {x: 0, y: 2, rotation: 0},
+    {x: 0, y: 1, rotation: 90},
+    {x: 0, y: 0, rotation: 90},
+    {x: 0, y: 1, rotation: 0}
+  ];
+  for (i = 0; i < 7; i++) {
+    segment = {
+      number: i,
+      layout: segmentLayoutMap[i],
+      component: this,
+      pin: this.pinMap[['a', 'b', 'c', 'd', 'e', 'f', 'g', 'DP'][i]]
+    };
+    this.segments.push(new Segment(segment));
+  }
+
+  this.decimalPoint = {
+    layout: {x: 1, y: 2}
+  };
+};
+LED.prototype.calculatePosition = function (selected, index, count) {
+  var selectedConstants = constants.selectedConstants(selected),
+      displayWidth, displayHeight, i, pin, pinDX, segmentWidth, segmentHeight, segment, pathCommands, endCapSize, p;
+
+  this.position = layout.calculateComponentRect(selected, index, count);
+
+  displayWidth = this.position.width * 0.8;
+  displayHeight = this.position.height * 0.9;
+
+  this.position.display = {
+    x: this.position.x + ((this.position.width - displayWidth) / 2),
+    y: this.position.y + ((this.position.height - displayHeight) / 2),
+    width: displayWidth,
+    height: displayHeight
+  };
+
+  // pins
+  pinDX = (this.position.display.width - (selectedConstants.PIN_WIDTH * 5)) / 6;
+  for (i = 0; i < this.pins.length; i++) {
+    pin = this.pins[i];
+    pin.x = this.position.display.x + pinDX + ((i % (this.pins.length / 2)) * (selectedConstants.PIN_WIDTH + pinDX));
+    pin.y = i < 5 ? this.position.display.y - selectedConstants.PIN_HEIGHT : this.position.display.y + this.position.display.height;
+    pin.cx = pin.x + (selectedConstants.PIN_WIDTH / 2);
+    pin.cy = pin.y + (selectedConstants.PIN_HEIGHT / 2);
+    pin.width = selectedConstants.PIN_WIDTH;
+    pin.height = selectedConstants.PIN_HEIGHT;
+    pin.labelSize = selectedConstants.PIC_FONT_SIZE;
+    pin.label.x = pin.x + (selectedConstants.PIN_WIDTH / 2);
+    pin.label.y = i < 5 ? this.position.display.y + (1.5 * selectedConstants.PIC_FONT_SIZE) : this.position.display.y + this.position.display.height - (0.75 * selectedConstants.PIC_FONT_SIZE);
+    pin.label.anchor = 'middle';
+  }
+
+  // segments
+  segmentWidth = this.position.display.width / 3;
+  segmentHeight = this.position.display.width / 12;
+  p = this.position.segments = {
+    x: this.position.display.x + ((this.position.display.width - segmentWidth) / 2),
+    y: this.position.display.y + ((this.position.display.height - (segmentWidth * 2)) / 2) - (segmentHeight / 2), // y is rotated to width = height
+    segmentWidth: segmentWidth,
+    segmentHeight: segmentHeight
+  };
+
+  endCapSize = segmentHeight / 2;
+  pathCommands = [
+    'M', p.x, ',', p.y + endCapSize, ' ',
+    'L', p.x + endCapSize, ',', p.y, ' ',
+    'L', p.x + segmentWidth - endCapSize, ',', p.y, ' ',
+    'L', p.x + segmentWidth, ',', p.y + endCapSize, ' ',
+    'L', p.x + segmentWidth - endCapSize, ',', p.y + segmentHeight, ' ',
+    'L', p.x + endCapSize, ',', p.y + segmentHeight, ' ',
+    'L', p.x, ',', p.y + endCapSize, ' '
+  ].join('');
+
+  for (i = 0; i < this.segments.length; i++) {
+    segment = this.segments[i];
+    segment.transform = ['translate(', segment.layout.x * segmentWidth, ',', segment.layout.y * segmentWidth, ')'].join('');
+    if (segment.layout.rotation) {
+      segment.transform = [segment.transform, ' rotate(', segment.layout.rotation, ' ', this.position.segments.x, ' ', this.position.segments.y + (segmentHeight / 2), ')'].join('');
+    }
+    segment.pathCommands = pathCommands;
+  }
+
+  this.decimalPoint.cx = this.position.segments.x + segmentWidth + segmentHeight + endCapSize;
+  this.decimalPoint.cy = this.position.segments.y + (2 * segmentWidth) + endCapSize;
+  this.decimalPoint.radius = endCapSize;
+};
+LED.prototype.reset = function () {
+  // nothing to do for LED
+};
+LED.prototype.resolveOutputValues = function () {
+  // nothing to do for LED
+};
+LED.prototype.getPinBitField = function () {
+  var bitfield = 0,
+      i;
+  for (i = 0; i < this.pins.length; i++) {
+    bitfield = bitfield | ((this.pins[i].value ? 1 : 0) << i);
+  }
+  return bitfield;
+};
+
+module.exports = LED;
