+++ app/src/models/pic/segment.js
@@ -1,0 +1,8 @@
+var Segment = function (options) {
+  this.index = options.index;
+  this.layout = options.layout;
+  this.component = options.component;
+  this.pin = options.pin;
+};
+
+module.exports = Segment;
