+++ app/src/views/pic/connector.js
@@ -1,0 +1,23 @@
+var ConnectorHoleView = React.createFactory(require('./connector-hole')),
+    svg = React.DOM.svg,
+    rect = React.DOM.rect;
+
+module.exports = React.createClass({
+  displayName: 'ConnectorView',
+
+  render: function () {
+    var position = this.props.connector.position,
+        holes = [],
+        hole, i;
+
+    for (i = 0; i < this.props.connector.holes.length; i++) {
+      hole = this.props.connector.holes[i];
+      holes.push(ConnectorHoleView({key: i, connector: this.props.connector, hole: hole, selected: this.props.selected, editable: this.props.editable, drawConnection: this.props.drawConnection, reportHover: this.props.reportHover}));
+    }
+
+    return svg({},
+      rect({x: position.x, y: position.y, width: position.width, height: position.height, fill: '#aaa'}),
+      holes
+    );
+  }
+});
