+++ app/src/views/pic/sidebar-chat.js
@@ -1,0 +1,122 @@
+var userController = require('../../controllers/user'),
+    ChatItems = React.createFactory(require('./chat-items')),
+    events = require('./events'),
+    div = React.DOM.div,
+    form = React.DOM.form,
+    textarea = React.DOM.textarea,
+    br = React.DOM.br,
+    button = React.DOM.button;
+
+module.exports = React.createClass({
+  displayName: 'SidebarChatView',
+
+  getInitialState: function() {
+    var items = [];
+
+    if (this.props.initialChatMessage) {
+      items.push({
+        prefix: 'Welcome!',
+        message: this.props.initialChatMessage
+      });
+    }
+
+    return {
+      items: items,
+      numExistingUsers: 0
+    };
+  },
+
+  getJoinedMessage: function (numExistingUsers) {
+    var slotsRemaining = 3 - numExistingUsers,
+        nums = ["zero", "one", "two", "three"],
+        cap = function (string) {
+          return string.charAt(0).toUpperCase() + string.slice(1);
+        },
+        message = " ";
+
+    if (slotsRemaining > 1) {
+      // One of three users is here
+      message += cap(nums[numExistingUsers]) + " of 3 users is here.";
+    } else if (slotsRemaining == 1) {
+      // Two of you are now here. One more to go before you can get started!
+      message += cap(nums[numExistingUsers]) + " of you are now here. One more to go before you can get started!";
+    } else {
+      message += "You're all here! Time to start this challenge.";
+    }
+
+    return message;
+  },
+
+  componentWillMount: function() {
+    var self = this;
+    userController.onGroupRefCreation(function() {
+      self.firebaseRef = userController.getFirebaseGroupRef().child("chat");
+      self.firebaseRef.orderByChild('time').on("child_added", function(dataSnapshot) {
+        var items = self.state.items.slice(0),
+            item = dataSnapshot.val(),
+            numExistingUsers = self.state.numExistingUsers;
+
+        if (item.type == "joined") {
+          numExistingUsers = Math.min(self.state.numExistingUsers + 1, 3);
+          item.message += self.getJoinedMessage(numExistingUsers);
+        }
+        else if (item.type == "left") {
+          numExistingUsers = Math.max(self.state.numExistingUsers - 1, 0);
+        }
+
+        if (numExistingUsers !== self.state.numExistingUsers) {
+          self.setState({numExistingUsers: numExistingUsers});
+        }
+
+        items.push(item);
+
+        self.setState({
+          items: items
+        });
+      }.bind(self));
+    });
+  },
+
+  componentWillUnmount: function() {
+    this.firebaseRef.off();
+  },
+
+  handleSubmit: function(e) {
+    var input = this.refs.text,
+        message = $.trim(input.value);
+
+    e.preventDefault();
+
+    if (message.length > 0) {
+      this.firebaseRef.push({
+        user: userController.getUsername(),
+        message: message,
+        time: Firebase.ServerValue.TIMESTAMP
+      });
+      input.value = '';
+      input.focus();
+      events.logEvent("Sent message", message);
+    }
+  },
+
+  listenForEnter: function (e) {
+    if (e.keyCode === 13) {
+      this.handleSubmit(e);
+    }
+  },
+
+  render: function () {
+    var style = this.props.demo ? {} : {top: 75};
+    return div({id: 'sidebar-chat', style: style},
+      div({id: 'sidebar-chat-title'}, 'Chat'),
+      ChatItems({items: this.state.items}),
+      div({className: 'sidebar-chat-input'},
+        form({onSubmit: this.handleSubmit},
+          textarea({ref: 'text', placeholder: 'Enter chat message here...', onKeyDown: this.listenForEnter}),
+          br({}),
+          button({onClick: this.handleSubmit}, 'Send Chat Message')
+        )
+      )
+    );
+  }
+});
